Modularize and implement Dropout mechanisms (#60)
- Add `drop_block` functionality inspired by DropBlock (Ghiasi et al., 2018) with extended configurability.
- Refactor `dropout` into a dedicated module for better management and separation of concerns.
- Relocate `dropout` and `drop_block` logic from `ops/drop.rs` to modular submodules.
- Enhance unit tests for `dropout` and `drop_block` operations to improve coverage and reliability.
