+++ crates/bunsen/src/kits/bimm/resnet/basic_block.rs
@@ -39,7 +39,6 @@ use crate::{
             drop_block::{
                 DropBlock2d,
                 DropBlock2dConfig,
-                DropBlockOptions,
             },
             drop_path::{
                 DropPath,
@@ -53,7 +52,10 @@ use crate::{
         ResNetDownsample,
         ResNetDownsampleConfig,
     },
-    ops::conv::stride_div_output_resolution,
+    ops::{
+        conv::stride_div_output_resolution,
+        drop::DropBlockOptions,
+    },
     support::{
         arrays::scalar_to_array,
         validators::expect_probability,
