+++ crates/bunsen/src/ops/drop/mod.rs
@@ -1,0 +1,9 @@
+//! Dropout Mechanisms
+
+mod drop_block;
+mod dropout_func;
+
+#[doc(inline)]
+pub use drop_block::*;
+#[doc(inline)]
+pub use dropout_func::*;
